% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binalysis-chunks.R
\name{sectionTitle,Binalysis-method}
\alias{sectionTitle,Binalysis-method}
\alias{load,Binalysis-method}
\alias{introduction,Binalysis-method}
\alias{parameters,Binalysis-method}
\alias{featureTable,Binalysis-method}
\alias{chromatograms,Binalysis-method}
\alias{fingerprints,Binalysis-method}
\alias{purityAndCentrality,Binalysis-method}
\alias{ticPlot,Binalysis-method}
\alias{rsdPlot,Binalysis-method}
\title{Report chunks for Binalysis class}
\usage{
\S4method{sectionTitle}{Binalysis}(x)

\S4method{load}{Binalysis}(x)

\S4method{introduction}{Binalysis}(x)

\S4method{parameters}{Binalysis}(x)

\S4method{featureTable}{Binalysis}(x)

\S4method{chromatograms}{Binalysis}(x)

\S4method{fingerprints}{Binalysis}(x)

\S4method{purityAndCentrality}{Binalysis}(x)

\S4method{ticPlot}{Binalysis}(x)

\S4method{rsdPlot}{Binalysis}(x)
}
\arguments{
\item{x}{S4 object of class Binalysis}
}
\value{
S4 object of class Chunk.
}
\description{
Report chunks for Binalysis S4 class
}
