% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{code,Chunk-method}
\alias{code,Chunk-method}
\alias{code<-,Chunk-method}
\alias{label,Chunk-method}
\alias{label<-,Chunk-method}
\alias{chunkOptions,Chunk-method}
\alias{chunkOptions<-,Chunk-method}
\alias{textAbove,Chunk-method}
\alias{textAbove<-,Chunk-method}
\alias{textBelow,Chunk-method}
\alias{textBelow<-,Chunk-method}
\title{Chunk get and set methods}
\usage{
\S4method{code}{Chunk}(x)

\S4method{code}{Chunk}(x, ...) <- value

\S4method{label}{Chunk}(x)

\S4method{label}{Chunk}(x) <- value

\S4method{chunkOptions}{Chunk}(x)

\S4method{chunkOptions}{Chunk}(x) <- value

\S4method{textAbove}{Chunk}(x)

\S4method{textAbove}{Chunk}(x) <- value

\S4method{textBelow}{Chunk}(x)

\S4method{textBelow}{Chunk}(x) <- value
}
\arguments{
\item{x}{S4 object of class Chunk}

\item{...}{R expressions to set}

\item{value}{value to set}
}
\description{
Get and set methods for Chunk S4 class.
}
