% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{chunk}
\alias{chunk}
\title{Genreate an Rmarkdown chunk}
\usage{
chunk(
  ...,
  label = "",
  chunk_options = list(),
  text_above = "",
  text_below = ""
)
}
\arguments{
\item{...}{R expressions from which to generate a chunk}

\item{label}{chunk label}

\item{chunk_options}{knitr chunk options}

\item{text_above}{optional text to be placed above the code chunk}

\item{text_below}{optional text to be placed below the code chunk}
}
\description{
Generate an Rmarkdown code chunk.
}
\examples{
chunk(a <- 1,
      b <- 2,
      a + b,
      label = 'example',
      chunk_options = list(eval = FALSE),
      text_above = 'Some example text above.',
      text_below = 'Some example text below.') 
}
