% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleFIEworkflowResults.R
\docType{data}
\name{exampleFIEworkflowResults}
\alias{exampleFIEworkflowResults}
\title{exampleFIEworkflowResults}
\format{An S4 object of class `Workflow`}
\usage{
exampleFIEworkflowResults
}
\description{
example FIE-HRMS workflow results from the example B. distachyon ecotype data in package \code{methaboData}.
}
\keyword{datasets}
