% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/parameters-access.R
\name{sessionInformation}
\alias{sessionInformation}
\alias{sessionInformation<-}
\alias{sessionInformation,ReportParameters-method}
\alias{sessionInformation<-,ReportParameters-method}
\title{Get and set report parameters}
\usage{
sessionInformation(x)

sessionInformation(x) <- value

\S4method{sessionInformation}{ReportParameters}(x)

\S4method{sessionInformation}{ReportParameters}(x) <- value
}
\arguments{
\item{x}{S4 object of class ReportParameters}

\item{value}{parameter value to set}
}
\description{
Get and set report paramters.
}
