% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/render.R
\name{renderReport}
\alias{renderReport}
\alias{renderReport,Report-method}
\title{Render a report}
\usage{
renderReport(x, open = TRUE, ...)

\S4method{renderReport}{Report}(x, open = TRUE, ...)
}
\arguments{
\item{x}{S4 object of class Report}

\item{open}{open report when generated}

\item{...}{arguments to pass to \code{rmarkdown::render}}
}
\description{
Render a report.
}
