% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report-chunks.R
\name{reportHeader,Report-method}
\alias{reportHeader,Report-method}
\alias{reportOptions,Report-method}
\alias{sessionInformationSection,Report-method}
\alias{reportFooter,Report-method}
\title{Report chunks}
\usage{
\S4method{reportHeader}{Report}(x)

\S4method{reportOptions}{Report}(x)

\S4method{sessionInformationSection}{Report}(x)

\S4method{reportFooter}{Report}(x)
}
\arguments{
\item{x}{S4 object of class Report}
}
\description{
General report chunks
}
