% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/report-access.R
\name{report}
\alias{report}
\alias{report<-}
\alias{report,Report-method}
\alias{report<-,Report-method}
\title{Get and set a report}
\usage{
report(x)

report(x) <- value

\S4method{report}{Report}(x)

\S4method{report}{Report}(x) <- value
}
\arguments{
\item{x}{S4 object of class Report}

\item{value}{list containing report sections to set}
}
\description{
Get and set a report in an S4 object of class Report.
}
