% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/report-access.R
\name{reportData}
\alias{reportData}
\alias{reportData<-}
\alias{reportData,Report-method}
\alias{reportData<-,Report-method}
\title{Get and set report data}
\usage{
reportData(x)

reportData(x) <- value

\S4method{reportData}{Report}(x)

\S4method{reportData}{Report}(x) <- value
}
\arguments{
\item{x}{S4 object of class Report}

\item{value}{list containing data objects to set}
}
\description{
Get and set data in an S4 object of class Report.
}
