% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/report-access.R
\name{reportObjects}
\alias{reportObjects}
\alias{reportObjects,Report-method}
\title{Object classes included in a report}
\usage{
reportObjects(x)

\S4method{reportObjects}{Report}(x)
}
\arguments{
\item{x}{S4 object of class Report}
}
\description{
Return object calsses in a report.
}
