% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ReportParameters-class}
\alias{ReportParameters-class}
\title{Report parameters class}
\description{
An S4 class to store report parameters.
}
\section{Slots}{

\describe{
\item{\code{path}}{report directory output path}

\item{\code{title}}{report title}

\item{\code{investigator}}{investigator name}

\item{\code{sessionInfo}}{logical, inclusion of session info}

\item{\code{toc}}{logical, inclusion of table of contents}

\item{\code{echo}}{logical, inclusion of code chunks}

\item{\code{cache}}{logical, cache data}

\item{\code{time}}{creation time}

\item{\code{output}}{output type}
}}

