% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportParameters.R
\name{reportParameters}
\alias{reportParameters}
\title{reportParameters}
\usage{
reportParameters(title, investigator, path = ".")
}
\arguments{
\item{title}{report title}

\item{investigator}{investigator name}

\item{path}{report output path}
}
\description{
Initiate report parameters.
}
