% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/report-access.R
\name{reportRMD}
\alias{reportRMD}
\alias{reportRMD,Report-method}
\title{Return a report .Rmd}
\usage{
reportRMD(x)

\S4method{reportRMD}{Report}(x)
}
\arguments{
\item{x}{S4 object of class Report}
}
\description{
Return a report markdown as a character string
}
