% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{rmd,Chunk-method}
\alias{rmd,Chunk-method}
\title{Chunk Rmarkdown}
\usage{
\S4method{rmd}{Chunk}(x)
}
\arguments{
\item{x}{S4 object of class Chunk}
}
\description{
Retrieve Rmarkdown from an object of class Chunk.
}
\examples{
x <- chunk(1 + 1)
rmd(x)
}
