% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show,Report-method}
\alias{show,Report-method}
\title{Report class show method}
\usage{
\S4method{show}{Report}(object)
}
\arguments{
\item{object}{S4 object of class Report}
}
\description{
Show method for Report class
}
