% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show,ReportParameters-method}
\alias{show,ReportParameters-method}
\title{ReportParameters class show method}
\usage{
\S4method{show}{ReportParameters}(object)
}
\arguments{
\item{object}{S4 object of class ReportParameters}
}
\description{
Show method for ReportParameters class
}
