% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binalysisMethods.R
\name{sectionTitle,Binalysis-method}
\alias{sectionTitle,Binalysis-method}
\alias{load,Binalysis-method}
\alias{introduction,Binalysis-method}
\alias{parameters,Binalysis-method}
\alias{featureTable,Binalysis-method}
\alias{chromatograms,Binalysis-method}
\alias{fingerprints,Binalysis-method}
\alias{purityAndCentrality,Binalysis-method}
\alias{ticPlot,Binalysis-method}
\alias{rsdPlot,Binalysis-method}
\title{Report chunks for Binalysis class}
\usage{
\S4method{sectionTitle}{Binalysis}(...)

\S4method{load}{Binalysis}(...)

\S4method{introduction}{Binalysis}(...)

\S4method{parameters}{Binalysis}(...)

\S4method{featureTable}{Binalysis}(...)

\S4method{chromatograms}{Binalysis}(...)

\S4method{fingerprints}{Binalysis}(...)

\S4method{purityAndCentrality}{Binalysis}(...)

\S4method{ticPlot}{Binalysis}(...)

\S4method{rsdPlot}{Binalysis}(...)
}
\arguments{
\item{...}{S4 object of class Binalysis}
}
\value{
S4 object of class Chunk.
}
\description{
Report chunks for Binalysis S4 class
}
