% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generateReport,ReportParameters-method}
\alias{generateReport,ReportParameters-method}
\title{Generate a report}
\usage{
\S4method{generateReport}{ReportParameters}(parameters, ...)
}
\arguments{
\item{parameters}{S4 object of class ReportParameters}

\item{...}{}
}
\description{
Generate a report skeleton.
}
