% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/parameters-access.R
\name{path}
\alias{path}
\alias{path<-}
\alias{title}
\alias{title<-}
\alias{investigator}
\alias{investigator<-}
\alias{sessionInformation}
\alias{sessionInformation<-}
\alias{toc}
\alias{toc<-}
\alias{echo}
\alias{echo<-}
\alias{cache}
\alias{cache<-}
\alias{creationDate}
\alias{creationDate<-}
\alias{output}
\alias{output<-}
\alias{path,ReportParameters-method}
\alias{path<-,ReportParameters-method}
\alias{title,ReportParameters-method}
\alias{title<-,ReportParameters-method}
\alias{investigator,ReportParameters-method}
\alias{investigator<-,ReportParameters-method}
\alias{sessionInformation,ReportParameters-method}
\alias{sessionInformation<-,ReportParameters-method}
\alias{toc,ReportParameters-method}
\alias{toc<-,ReportParameters-method}
\alias{echo,ReportParameters-method}
\alias{echo<-,ReportParameters-method}
\alias{cache,ReportParameters-method}
\alias{cache<-,ReportParameters-method}
\alias{creationDate,ReportParameters-method}
\alias{creationDate<-,ReportParameters-method}
\alias{output,ReportParameters-method}
\alias{output<-,ReportParameters-method}
\title{Get and set report parameters}
\usage{
path(x)

path(x) <- value

title(x)

title(x) <- value

investigator(x)

investigator(x) <- value

sessionInformation(x)

sessionInformation(x) <- value

toc(x)

toc(x) <- value

echo(x)

echo(x) <- value

cache(x)

cache(x) <- value

creationDate(x)

creationDate(x) <- value

output(x)

output(x) <- value

\S4method{path}{ReportParameters}(x)

\S4method{path}{ReportParameters}(x) <- value

\S4method{title}{ReportParameters}(x)

\S4method{title}{ReportParameters}(x) <- value

\S4method{investigator}{ReportParameters}(x)

\S4method{investigator}{ReportParameters}(x) <- value

\S4method{sessionInformation}{ReportParameters}(x)

\S4method{sessionInformation}{ReportParameters}(x) <- value

\S4method{toc}{ReportParameters}(x)

\S4method{toc}{ReportParameters}(x) <- value

\S4method{echo}{ReportParameters}(x)

\S4method{echo}{ReportParameters}(x) <- value

\S4method{cache}{ReportParameters}(x)

\S4method{cache}{ReportParameters}(x) <- value

\S4method{creationDate}{ReportParameters}(x)

\S4method{creationDate}{ReportParameters}(x) <- value

\S4method{output}{ReportParameters}(x)

\S4method{output}{ReportParameters}(x) <- value
}
\arguments{
\item{x}{S4 object of class ReportParameters}

\item{value}{parameter value to set}
}
\description{
Get and set report paramters.
}
