% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{report}
\usage{
report(analysis, parameters)
}
\arguments{
\item{analysis}{object of class Binalysis from which to genrate a report}

\item{parameters}{object of class ReportParameters}
}
\description{
generate a report from a Binalysis analysis.
}
\examples{
\dontrun{
## For just a Binalysis object
data("exampleFIEworkflowResults")
analysis <- exampleFIEworkflowResults \%>\%
 resultsProcessing()
p <- reportParameters('B. distachyon ecotype comparison','Steve French')
report(analysis,p)
}
}
