% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{report}
\usage{
report(analysis, parameters)
}
\arguments{
\item{analysis}{object of class Binalysis from which to genrate a report}

\item{parameters}{object of class ReportParameters}
}
\description{
generate a report from a Binalysis analysis.
}
\examples{
\dontrun{
library(metaboWorkflows)

files <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes') 
info <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')

wp <- workflowParameters('FIE-HRMS fingerprinting',files = files)
flags(wp) <- flags(wp)[-c(5,7)]
analysis <- workflow(files, info, wp)

## For just a Binalysis object
bd <- analysis \%>\%
 resultsProcessing()
p <- reportParameters('ExampleData','Steve French')
report(bd,p)

## For Analysis object
ad <- analysis \%>\%
 resultsAnalysis()
p <- reportParameters('ExampleData','Steve French')
report(ad,p) 

## For Assignment object
as <- analysis \%>\%
 resultsAnnotation()
p <- reportParameters('ExampleData','Steve French')
report(as,p) 

## For workflow object
p <- reportParameters('ExampleData','Steve French')
report(analysis,p) 

}
}
