% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{report}
\usage{
report(analysis, parameters)
}
\description{
report
}
\examples{

## For just a Binalysis object
data("exampleFIEworkflowResults")
analysis <- analysis \%>\%
 resultsProcessing()
p <- reportParameters('test','Steve French')
report(analysis,p)

}
