% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/report.R
\name{report}
\alias{report}
\alias{report,Binalysis-method}
\alias{report,MetaboProfile-method}
\alias{report,Analysis-method}
\alias{report,Assignment-method}
\alias{report,Workflow-method}
\title{report}
\usage{
report(analysis, parameters)

\S4method{report}{Binalysis}(analysis, parameters)

\S4method{report}{MetaboProfile}(analysis, parameters)

\S4method{report}{Analysis}(analysis, parameters)

\S4method{report}{Assignment}(analysis, parameters)

\S4method{report}{Workflow}(analysis, parameters)
}
\arguments{
\item{analysis}{S4 object of class Binalysis, MetaboProfile, Analysis, Assignment or Workflow from which to genrate a report}

\item{parameters}{object of class ReportParameters}
}
\description{
Automated, dynamic report generation for a metabolomic analysis.
}
\examples{
\dontrun{
library(metaboWorkflows)

fp <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes') 
si <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')

wp <- workflowParameters('FIE-HRMS fingerprinting',fp,si,breaks = FALSE)

analysis <- workflow(wp)

## For a Binalysis object
bd <- analysis \%>\%
 resultsProcessing()
p <- reportParameters('ExampleData','Steve French')
report(bd,p)

## For an Analysis object
ad <- analysis \%>\%
 resultsAnalysis()
p <- reportParameters('ExampleData','Steve French')
report(ad,p) 

## For an Assignment object
as <- analysis \%>\%
 resultsAnnotation()
p <- reportParameters('ExampleData','Steve French')
report(as,p) 

## For a Workflow object
p <- reportParameters('ExampleData','Steve French')
report(analysis,p) 

}
}
