% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{report}
\usage{
report(analysis, parameters)
}
\arguments{
\item{analysis}{object of class Binalysis from which to genrate a report}

\item{parameters}{object of class ReportParameters}
}
\description{
generate a report from a Binalysis analysis.
}
\examples{
\dontrun{
library(metaboWorkflows)
library(magrittr)
data("exampleFIEworkflowResults")

## For just a Binalysis object
analysis <- exampleFIEworkflowResults \%>\%
 resultsProcessing()
p <- reportParameters('ExampleData','Steve French')
report(analysis,p)

## For Analysis object
analysis <- exampleFIEworkflowResults \%>\%
 resultsAnalysis()
p <- reportParameters('ExampleData','Steve French')
report(analysis,p) 

## For Assignment object
analysis <- exampleFIEworkflowResults \%>\%
 resultsAnnotation()
p <- reportParameters('ExampleData','Steve French')
report(analysis,p) 

## For workflow object
p <- reportParameters('ExampleData','Steve French')
report(exampleFIEworkflowResults,p) 

}
}
