% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportParameters.R
\name{reportParameters}
\alias{reportParameters}
\title{Report parameters}
\usage{
reportParameters(
  title,
  investigator,
  sessionInfo = TRUE,
  toc = TRUE,
  echo = FALSE,
  cache = FALSE,
  path = "."
)
}
\arguments{
\item{title}{report title}

\item{investigator}{investigator name}

\item{sessionInfo}{TRUE/FALSE, include session info}

\item{toc}{TRUE/FALSE, include a table of contents}

\item{echo}{TRUE/FALSE, echo code chunks}

\item{cache}{TRUE/FALSE, cache report code chunk data}

\item{path}{report output path}
}
\description{
Initiate report parameters.
}
\seealso{
\code{\link{report}}
}
