% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportParameters.R
\name{reportParameters}
\alias{reportParameters}
\title{reportParameters}
\usage{
reportParameters(
  title,
  investigator,
  sessionInfo = TRUE,
  toc = TRUE,
  path = "."
)
}
\arguments{
\item{title}{report title}

\item{investigator}{investigator name}

\item{sessionInfo}{TRUE/FALSE, should session info included?}

\item{toc}{TRUE/FALSE, should a table of contents be included}

\item{path}{report output path}
}
\description{
Initiate report parameters.
}
